added genre feature and tested correctly

+++ src/controllers/helper.js
@@ -1,15 +1,17 @@
-const { Book, Reader, Author } = require("../models");
+const { Book, Reader, Author, Genre } = require("../models");
 const getModel = (model) => {
   const models = {
     reader: Reader,
     book: Book,
     author: Author,
+    genre: Genre,
   };
   return models[model];
 };
 const readerError = { error: "The reader could not be found." };
 const bookError = { error: "The book could not be found." };
 const authorError = { error: "The author could not be found." };
+const genreError = { error: "The genre could not be found." };
 
 const removePassword = (obj) => {
   if (obj.hasOwnProperty("password")) {
@@ -40,6 +42,9 @@ const findById = async (model, req, res) => {
   if (!thisItem && model == "author") {
     return res.status(404).send(authorError);
   }
+  if (!thisItem && model == "genre") {
+    return res.status(404).send(genreError);
+  }
   res.status(200).send(removePassword(thisItem.dataValues));
 };
 const update = async (model, req, res) => {
@@ -52,6 +57,9 @@ const update = async (model, req, res) => {
   }
   if (!thisItem && model == "author") {
     return res.status(404).send(authorError);
+  }
+  if (!thisItem && model == "genre") {
+    return res.status(404).send(genreError);
   }
   await getModel(model).update(req.body, {
     where: { id: req.params.id },
@@ -70,6 +78,10 @@ const remove = async (model, req, res) => {
   if (!thisItem && model == "author") {
     return res.status(404).send(authorError);
   }
+  if (!thisItem && model == "genre") {
+    return res.status(404).send(genreError);
+  }
+
   await getModel(model).destroy({
     where: { id: req.params.id },
   });
